Separate dev/prod and dockerise prod
- Backend now uses dev and prod profiles, which can be triggered thanks
to the SPRING_PROFILES_ACTIVE env variable.
- Frontend now uses dev and prod env files, which can be triggered
thanks to ng build configuration and a switch specified in angular.json
file.
- Dev mode run is done the same as before but prod is made as a
production ready deployment on docker thanks to a docker compose, also
including postgresql in that case, with the right driver used in the
backend if it's prod instead of dev.