Add RabbitMQ and separate maven profiles
- Added a Hello World like implementation of RabbitMQ flow, just on the
create list for now, and a receiver will show the message in the output.
This will of course improve with a more meaningful setup.
- To make the app running in dev mode, it requires a RabbitMQ instance
that it can use therefore an internal compose is created to create a MQ
container. Of course with some tweaking, it should be possible to not
use docker, but this is for simplicity.
As for prod, this is simply added in the main compose.
- In order to cleanly separate dependencies and what's needed, it's
using maven profile to only include certain deps, such as the spring
docker compose one for dev.
- Made sure the tests ignore the listener for now.