Add Journal structure and MQ receivers
Goal is to keep track of actions as log form, but the specificity here
is that a RabbitMQ publish/subscribe is integrated in order to
asynchronously manage these action messages:
- A journal table is created to persist entries regarding TodoList
actions, currently only Create List.
- Moreover, a cache mechanism is integrated to only contain the latest
Journal additions with the goal of being polled by the frontend
afterward, this cache will be cleared whenever the Journal controller
endpoints are called.